+++ stickyuploads/__init__.py
@@ -1,3 +1,3 @@
 "Enhanced file input widget for Django which uploads the file in the background and retains value on form errors."
 
-__version__ = '0.6.1'
+__version__ = '1.0.0'