+++ stickyuploads/tests/base.py
@@ -3,12 +3,21 @@
 import shutil
 import tempfile
 
+from django.db import models
+
+try:
+    from django.utils.six import string_types, text_type
+except ImportError:
+    # Too new, no six, must be python 3
+    string_types = [str]
+    text_type = str
+
 from mock import Mock
 
 
 class MockStorage(Mock):
 
-    def save(self, name, upload, max_length=None):
+    def save(self, name, content, max_length=None):
         return name
 
     def url(self, name):
@@ -16,6 +25,9 @@ class MockStorage(Mock):
 
     def get_valid_name(self, name):
         return name
+
+    def generate_filename(self, thing):
+        return text_type(thing)
 
 
 mockstorage = MockStorage()
@@ -33,3 +45,8 @@ class TempFileMixin(object):
     def tearDown(self):
         super(TempFileMixin, self).tearDown()
         shutil.rmtree(self.temp_dir, ignore_errors=True)
+
+
+class Example(models.Model):
+    name = models.CharField(max_length=100)
+    upload = models.FileField(storage=mockstorage, upload_to='test/')